🌳 Refactor: useHomePage hook 생성 및 적용
